refactor: navigation component updated
