+++ src/app/globals.css
@@ -1,36 +1,67 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-/* :root {
-  --foreground-rgb: 0, 0, 0;
-  --background-start-rgb: 214, 219, 220;
-  --background-end-rgb: 255, 255, 255;
+/*
+  Josh's Custom CSS Reset
+  https://www.joshwcomeau.com/css/custom-css-reset/
+*/
+*,
+*::before,
+*::after {
+  box-sizing: border-box;
 }
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --foreground-rgb: 255, 255, 255;
-    --background-start-rgb: 0, 0, 0;
-    --background-end-rgb: 0, 0, 0;
-  }
-} */
-
+* {
+  margin: 0;
+}
 html,
+body,
+#root {
+  height: 100%;
+}
 body {
-  font-family: var(--font-sansProMedium);
+  line-height: 1.5;
+  -webkit-font-smoothing: antialiased;
 }
-
-h1,
-h2 {
-  font-family: var(--font-butlerBlack);
+img,
+picture,
+video,
+canvas,
+svg {
+  display: block;
+  max-width: 100%;
 }
-
+input,
+button,
+textarea,
+select {
+  font: inherit;
+}
 p,
-a {
-  font-size: 1rem;
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  overflow-wrap: break-word;
+}
+#root,
+#__next {
+  isolation: isolate;
 }
 
+:root {
+  --transparent: transparent;
+  --white: #ffffff;
+  --black: #000000;
+  --gray: #9192a4;
+  --seasalt: #fafafa;
+  --night: #232323;
+  --pink: #ee0979;
+  --pumpkin: #ff6a00;
+  --red: #f42f48;
+
+  --black-white: var(--black);
+}
+
+/* Custom Fonts */
 .font-sansProExtraBold {
   font-family: var(--font-sansProExtraBold);
 }
@@ -48,4 +79,35 @@ a {
 }
 .font-butlerBlack {
   font-family: var(--font-butlerBlack);
+}
+
+[data-theme="dark"] {
+  --black-white: var(--white);
+}
+
+/* Custom global styles */
+html,
+body {
+  font-family: var(--font-sansProMedium);
+}
+
+h1,
+h2 {
+  font-family: var(--font-butlerBlack);
+}
+
+p,
+a {
+  font-size: 1rem;
+}
+
+a {
+  text-decoration: none;
+}
+
+.two-trans {
+  -webkit-transition: ease-in-out 0.2s;
+  -moz-transition: ease-in-out 0.2s;
+  -o-transition: ease-in-out 0.2s;
+  transition: ease-in-out 0.2s;
 }
