+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import Footer from "@/components/footer";
-import { Header } from "@/components/header";
+import { Header } from "@/components/header/header";
 import { Providers } from "../lib/providers/ThemeProvider/providers";
 import {
   butlerBlack,
