+++ src/components/header/navigation/Navigation.module.css
@@ -1,0 +1,34 @@
+.nav {
+  color: var(--gray);
+  text-transform: uppercase;
+}
+
+.navLink {
+  margin: 0 1.875rem;
+  letter-spacing: 0.2em;
+  font-size: 0.75rem;
+  color: var(--gray);
+}
+
+.navLink:hover {
+  color: rgb(238, 9, 121);
+  color: -moz-linear-gradient(
+    180deg,
+    rgba(238, 9, 121, 1) 0%,
+    rgba(255, 106, 0, 1) 100%
+  );
+  color: -webkit-linear-gradient(
+    180deg,
+    rgba(238, 9, 121, 1) 0%,
+    rgba(255, 106, 0, 1) 100%
+  );
+  color: linear-gradient(
+    180deg,
+    rgba(238, 9, 121, 1) 0%,
+    rgba(255, 106, 0, 1) 100%
+  );
+}
+
+.active {
+  color: var(--black-white);
+}
