+++ src/lib/providers/ThemeProvider/providers.tsx
@@ -1,7 +1,7 @@
-'use client';
+"use client";
 
-import { ThemeProvider } from 'next-themes';
+import { ThemeProvider } from "next-themes";
 
 export function Providers({ children }: { children: any }) {
-  return <ThemeProvider attribute="class">{children}</ThemeProvider>;
+  return <ThemeProvider>{children}</ThemeProvider>;
 }
